refactor(UnauthenticatedLayout): move component to its own directory
