+++ src/components/router/Router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 
-import UnauthenticatedLayout from '~components/layouts/UnauthenticatedLayout'
+import UnauthenticatedLayout from '~/components/layouts/UnauthenticatedLayout/UnauthenticatedLayout'
 import Login from '~/pages/login/Login'
 import ValidateAuthentication from '~components/auth/Authentication'
 import Home from '~/pages/home/Home'
